+++ test/inspec/author-publish-dispatcher_spec.rb
@@ -10,8 +10,8 @@ aem_base ||= '/opt'
 aem_port = @hiera.lookup('author::aem_port', nil, @scope)
 aem_port ||= '4502'
 
-author_data_volume_mount_point = @hiera.lookup('aem_curator::install_author::data_volume_mount_point', nil, @scope)
-author_data_volume_mount_point ||= '/mnt/ebs1'
+aem_author_keystore_path = @hiera.lookup('aem_curator::install_author::aem_keystore_path', nil, @scope)
+aem_author_keystore_path ||= '/etc/ssl/aem-author/author.ks'
 
 ### SSM paramter store lookup is only supported for hiera5
 # aem_author_keystore_password = @hiera.lookup('aem_curator::install_author::aem_keystore_password', nil, @scope)
@@ -57,8 +57,16 @@ end
 #   it { should_not match(/changeit/) }
 # end
 
-# Test if default keystore password is not changeit
-describe command("keytool -list -keystore #{author_data_volume_mount_point}/author/crx-quickstart/ssl/aem.ks -alias cqse -storepass changeit") do
+describe file(aem_author_keystore_path) do
+  it { should be_file }
+  it { should exist }
+  its('mode') { should cmp '00640' }
+  it { should be_owned_by 'aem-author' }
+  it { should be_grouped_into 'aem-author' }
+end
+
+# Test if default keystore password is changeit
+describe command("keytool -list -keystore #{aem_author_keystore_path} -alias cqse -storepass changeit") do
   its('exit_status') { should_not eq 0 }
 end
 
@@ -84,8 +92,8 @@ aem_base ||= '/opt'
 aem_port = @hiera.lookup('publish::aem_port', nil, @scope)
 aem_port ||= '4503'
 
-publish_data_volume_mount_point = @hiera.lookup('aem_curator::install_publish::data_volume_mount_point', nil, @scope)
-publish_data_volume_mount_point ||= '/mnt/ebs2'
+aem_publish_keystore_path = @hiera.lookup('aem_curator::install_publish::aem_keystore_path', nil, @scope)
+aem_publish_keystore_path ||= '/etc/ssl/aem-publish/publish.ks'
 
 ### SSM paramter store lookup is only supported for hiera5
 # aem_publish_keystore_password = @hiera.lookup('aem_curator::install_publish::aem_keystore_password', nil, @scope)
@@ -131,8 +139,16 @@ end
 #   it { should_not match(/changeit/) }
 # end
 
+describe file(aem_publish_keystore_path) do
+  it { should be_file }
+  it { should exist }
+  its('mode') { should cmp '00640' }
+  it { should be_owned_by 'aem-publish' }
+  it { should be_grouped_into 'aem-publish' }
+end
+
 # Test if default keystore password is changeit
-describe command("keytool -list -keystore #{publish_data_volume_mount_point}/publish/crx-quickstart/ssl/aem.ks -alias cqse -storepass changeit") do
+describe command("keytool -list -keystore #{aem_publish_keystore_path} -alias cqse -storepass changeit") do
   its('exit_status') { should_not eq 0 }
 end